add subs stock

+++ lib/domain/entities/stock_model.dart
@@ -5,7 +5,7 @@ class StockModel {
   late String numbers;
   late String price;
   late String labelCompany;
-  late Float interest;
+  late double interest;
 
   StockModel(this.nameCompany, this.numbers, this.price, this.labelCompany , this.interest);
 }
